backend/pyproject.toml
[project]
name = "sku-ops-backend"
version = "0.1.0"
description = "SKU Operations — backend API"
authors = [{ name = "nooz" }]
requires-python = ">=3.13,<3.14"
readme = "README.md"
license = "MIT"

dependencies = [
    # ── Core ──
    "fastapi>=0.110.1",
    "uvicorn>=0.25.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.6.4",
    "email-validator>=2.2.0",
    "python-multipart>=0.0.9",
    # ── Auth ──
    "pyjwt>=2.12.0",
    "bcrypt==4.1.3",
    "passlib>=1.7.4",
    # ── Database ──
    "asyncpg>=0.29.0",
    # ── AI / LLM ──
    "pydantic-ai>=1.0.0",
    "pydantic-ai-slim[anthropic,openrouter]>=1.0.0",
    "openai>=1.40.0",
    "tiktoken>=0.7.0",
    "rank-bm25>=0.2.2",
    "numpy>=1.26.0",
    # ── Document processing ──
    "Pillow>=10.0.0",
    "pytesseract>=0.3.10",
    "pypdf>=6.9.1",
    "pdf2image>=1.17.0",
    # ── HTTP / OAuth ──
    "requests>=2.31.0",
    "httpx>=0.27.0",
    "requests-oauthlib>=2.0.0",
    "cryptography>=42.0.8",
    # ── Config / Serialization ──
    "PyYAML>=6.0",
    "tzdata>=2024.2",
    # ── Security constraints ──
    "pyasn1>=0.6.3", # CVE-2026-30922 — transitive via google-genai→google-auth→rsa
    # ── Production hardening ──
    "python-json-logger>=2.0.7",
    "prometheus-client>=0.20.0",
    "sentry-sdk[fastapi]>=1.40.0",
    "redis>=7.3.0",
    "sqlmodel>=0.0.37",
    "sqlalchemy>=2.0.48",
    "supabase>=2.28.2",
    "greenlet>=3.3.2",
    "uuid6>=2025.0.1",
]

[dependency-groups]
dev = [
    "pytest>=8.0,<10",
    "pytest-asyncio>=0.23,<1.4",
    "pytest-cov>=6.0",
    "pytest-timeout>=2.3",
    "ruff>=0.15.0",
    "commitizen>=4.12.1",
    "cz-conventional-gitmoji>=0.7.0",
    "rich>=14.0.0",
    "pip-audit>=2.7",
    "openpyxl>=3.1.5",
]

# =============================================================================
# PYTEST
# =============================================================================
[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
testpaths = ["tests"]
python_files = ["test_*.py"]
pythonpath = [".", ".."]
addopts = ["--import-mode=importlib"]

# =============================================================================
# RUFF — linter + formatter
# =============================================================================
[tool.ruff]
fix = true
line-length = 100
src = ["."]
target-version = "py313"

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = [
    "A",      # flake8-builtins
    "ARG",    # flake8-unused-arguments
    "ASYNC",  # flake8-async
    "B",      # flake8-bugbear
    "BLE",    # flake8-blind-except
    "C4",     # flake8-comprehensions
    "DTZ",    # flake8-datetimez
    "E",      # pycodestyle errors
    "ERA",    # eradicate (commented-out code)
    "F",      # pyflakes
    "FLY",    # flynt
    "FURB",   # refurb
    "G",      # flake8-logging-format
    "I",      # isort
    "ICN",    # flake8-import-conventions
    "INP",    # flake8-no-pep420
    "INT",    # flake8-gettext
    "ISC",    # flake8-implicit-str-concat
    "LOG",    # flake8-logging
    "N",      # pep8-naming
    "PERF",   # perflint
    "PGH",    # pygrep-hooks
    "PIE",    # flake8-pie
    "PLC",    # pylint conventions
    "PLE",    # pylint errors
    "PLW",    # pylint warnings
    "PT",     # flake8-pytest-style
    "PYI",    # flake8-pyi
    "Q",      # flake8-quotes
    "RET",    # flake8-return
    "RSE",    # flake8-raise
    "RUF",    # ruff-specific
    "S",      # bandit (security)
    "SIM",    # flake8-simplify
    "SLF",    # flake8-self
    "SLOT",   # flake8-slots
    "T10",    # flake8-debugger
    "T20",    # flake8-print
    "TC",     # flake8-type-checking
    "TID",    # flake8-tidy-imports
    "TRY",    # tryceratops
    "UP",     # pyupgrade
    "W",      # pycodestyle warnings
    "YTT",    # flake8-2020
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.pycodestyle]
max-doc-length = 100

[tool.ruff.lint.pydocstyle]
convention = "numpy"

# =============================================================================
# COMMITIZEN — conventional commit message validation (pre-commit hook)
# =============================================================================
[tool.commitizen]
name = "cz_gitmoji"

[tool.setuptools.packages.find]
include = [
    "shared*",
    "identity*",
    "catalog*",
    "inventory*",
    "operations*",
    "purchasing*",
    "finance*",
    "documents*",
    "assistant*",
    "reports*",
    "jobs*",
    "api*",
]
